# NSwag/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:51219",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/Ingredient/{id}": {
      "get": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "Ingredient_Get",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/IngredientDto"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredient"
        ],
        "operationId": "Ingredient_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Order/customer/{customerId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "Order_GetAllOrderForCustomer",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "customerId",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CustomerViewOrderDto"
              }
            }
          }
        }
      }
    },
...
    "/api/Order/{orderId}": {
      "delete": {
        "tags": [
          "Order"
        ],
        "operationId": "Order_DeleteOrder",
        "parameters": [
          {
            "type": "integer",
            "name": "orderId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "IngredientDto": {
      "type": "object",
      "required": [
        "id",
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "CustomerViewOrderDto": {
      "type": "object",
      "required": [
        "id",
        "sandwiches",
        "price",
        "time",
        "state",
        "sandwichModifications"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "sandwiches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SandwichDto"
          }
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "time": {
          "type": "string",
          "format": "date-time"
        },
        "state": {
          "$ref": "#/definitions/OrderState"
        },
        "sandwichModifications": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/IngredientModificationDto"
            }
          }
        }
      }
    },
    "SandwichDto": {
      "type": "object",
      "required": [
        "id",
        "name",
        "price",
        "ingredients"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "ingredients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IngredientDto"
          }
        }
      }
    },
    "OrderState": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "OrderPlaced",
        "InTheOven",
        "ToBeDelivered",
        "Delivered",
        "Rejected"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4
      ]
    },
    "IngredientModificationDto": {
      "type": "object",
      "required": [
        "ingredientId",
        "sandwichId",
        "modification"
      ],
      "properties": {
        "ingredientId": {
          "type": "integer",
          "format": "int32"
        },
        "sandwichId": {
          "type": "integer",
          "format": "int32"
        },
        "modification": {
          "$ref": "#/definitions/Modifier"
        }
      }
    },
    "Modifier": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Add",
        "Remove"
      ],
      "enum": [
        0,
        1
      ]
    },
...
  }
}